Add list of 'ustanove', 'odjeli' and 'pregledi' to admin user page together with some actions
Admin can now see a full list of users, 'ustanova', 'odjel' and 'pregledi'. Actions like activate,
deactivate and delete are added to 'users' and 'ustanove'.
List of 'ustanove' in 'pretrazi' are now fetched only if they are marked as active.
Some extra methods for CRUD manipulation data are also added.