Adding all implementation of date which can be found in 'pregled' and 'pregledPacijenta'
On start of application all dates of 'pregledi' are set to day of launching application if
they are null. On user page, 'pregledPacijenta' is set to 'expired' if date of that pregled
is below current day. On 'pregled' page, user will not be able to sign on that pregled if
he already is signed and date of signing is bigger then that day and below the next date
of that pregled.