Change active property of UserInfo to status
Change property to status because there will be 3 option of user status. First is Incomplete,
when user just registrate but not fill others property like oib. Second is active when all data
of user are present. Only with this status user can sign  on pregledi. Thirs status is deactivated
status, when admin deactivate user for some reason.